Add missing use statements in test file
The design of Perl 6 requires that use statements only have effect in a
lexical scope. Otherwise it's impossible to use different (versions of)
modules sharing a name in different parts of a program.

This means essentially that we need to explicitly use all modules we need
in a scope.

See https://gist.github.com/niner/70f7b46eefb7e22af78d896bea11efeb for
further information.